Change from GDI custom window stuff to console
This project's goal was to develop a simple graphics engine to use with
physics simulations, not to make a custom GDI window interface. I've now
switched focus to just simply using the console window to draw.

The current code in SGWindow.cpp is from StackOverFlow:
https://stackoverflow.com/questions/12378642/c-pixels-in-console-window

Also, I've removed the WinMain() method, because it is not necessary to
build a window.