Refactor main() to main.cpp and add SGWindow class
The main() method should be the entry point of the program, which should
initialize the SGEngine. The SGEngine should be the class that stores a
reference to SGWindow, as the engine will be the one that obtains the
graphics and sends it to the SGWindow. This will allow a physics engine
to display images at regular time intervals too.