Temporary fix on rendering image
The SDL_LockTexture() and SDL_UnlockTexture() has been acting strangely
in the code. For the time being, I've switched over to the inefficient
texturing method, SDL_UpdateTexture(). I have also changed some of the
indexing for the pixel/window size. I didn't realize that some of the
SDL methods required bytes and/or pixels.